+++ DentalSystem/DentalSystem.Entities/Models/Payment.cs
@@ -8,7 +8,7 @@ namespace DentalSystem.Entities.Models
     {
         [Key] public int PaymentId { get; set; }
         [ForeignKey("AccountsReceivable")] public int AccountsReceivableId { get; set; }
-        public int AmountPaid { get; set; }
+   public decimal AmountPaid { get; set; }
         public DateTime PaymentDate { get; set; }
         public DateTime? DeletedOn { get; set; }
         public AccountsReceivable AccountsReceivable { get; set; }
